fixed import circle with rtsp server struct
