+++ server/apiServer/aes256Handler.go
@@ -1,4 +1,4 @@
-package api
+package apiServer
 
 import (
 	"crypto/aes"
