+++ server/apiServer/escort/paths/info/response.go
