+++ server/apiServer/flagship/paths/register/body.go
