+++ server/apiServer/flagship/paths/register/response.go
