+++ server/rtspServer/flagship/handler.go
@@ -1,0 +1,26 @@
+package flagship
+
+import (
+	"github.com/bluenviron/gortsplib/v4"
+	"strconv"
+	"time"
+)
+
+var ServerHandler *FlagshipHandler
+
+func InitRtspServer(rtspPortInt int) *FlagshipHandler {
+	rtspPort := ":" + strconv.Itoa(rtspPortInt)
+
+	ServerHandler := &FlagshipHandler{}
+	ServerHandler.Server = &gortsplib.Server{
+		RTSPAddress:              rtspPort,
+		ReadTimeout:              10 * time.Second,
+		WriteTimeout:             10 * time.Second,
+		WriteQueueSize:           512,
+		MaxPacketSize:            1472,
+		DisableRTCPSenderReports: false,
+		Handler:                  ServerHandler,
+	}
+
+	return ServerHandler
+}
